Minor refactoring for project structure
