+++ cmd/kube/apiextensions/v1/register.go
